Remove need for tput(1)
This should improve performance of AE. It should _somewhat_ be more
portable, too. However, the portability is questionable, because you'd
now _need_ to be on the targeted terminal type. The plus side is that
the targeted terminal type is what almost everyone in Linux uses, such
as:

* Alacritty
* Kitty
* Konsole
* Terminology
* GNOME Terminal
* XFCE4 Terminal

Now the only dependencies are coreutils and file(1), the latter of
which I would love to ditch, but I'm not quite sure how.