Don't override `TIMEFORMAT` & fix `-c` feature
ZSH, BASH, 'sh', and now specifically DASH will now properly work with
shebangs which do and do not use env(1). Before this commit, shebangs
pointing to DASH, explicitly, would've prevented `-c` (checking -
feature) to fail, and the use of env(1) would've also caused any of
them to not work properly with that feature.

The `TIMEFORMAT` environment variable (for BASH's `time` builtin) is
no longer overridden, so the user can provide their own, if preferred.
The usual value AE has used for ages is _still_ used, but only if that
variable is not empty.