Correct once mode (cursor) & address STDIN issue
The STDIN issue refers to when the target FILE is waiting for a
response from STDIN, such as by a prompt (IE: `read` in shells), but
when the user sends the interrupt signal (`^C`, SIGINT, or 130), in
order to bail from the prompt and exit AE, a newline would not be
added, potentially and likely breaking the user's prompt display. It's
a minor issue, but it's annoying.

This is only relevant for the once only mode, where AE executes the
FILE once (for testing, mainly). This commit should address this
problem, but the caveat is that the prompt or displayed line at the
time of awaiting STDIN is removed upon exit. Usually, I would say
that's unacceptable, but given it's likely just a prompt, I think it's
_probably_ fine. If anyone disagrees, let me know!

A comment was placed next to the `Progrm` variable assigning, because
I almost changed it to 'autoexec', which would've broken the `Usage()`
output, and potentially other things, if the user runs AE via the `ae`
link provided by the installer script. Thus, it is required.

Prior to this commit, once only mode didn't restore the cursor, so it
stayed hidden, which was bad anyway, since the cursor shouldn't have
been hidden in that mode in the first place. This commit addresses that
problem.